Source Files/Mocks/generator/requirements.txt
torch
torchvision
torchaudio
diffusers[training]
#smartcrop
#numpy
#git+https://github.com/huggingface/transformers.git
transformers
#layoutparser[layoutmodels]
#git+https://github.com/facebookresearch/detectron2.git@ff53992b1985b63bd3262b5a36167098e3dada02#egg=detectron2
#git+https://github.com/facebookresearch/detectron2.git@main#egg=detectron2
ultralytics
pytesseract
python-doctr[torch]
#git+https://github.com/pidahbus/deep-image-orientation-angle-detection.git@main
datasets
#pytorch_merge
tensorflow
#keras
#keras==2.15.0
#tf-keras
numpy<2
tf2onnx
torchvision
git+https://github.com/sberbank-ai/Real-ESRGAN
pyyaml>=5.3.1
requests
#jxlpy>=0.9.5
accelerate
diffusers
Pillow
safetensors
markdown
beautifulsoup4
toml
python-frontmatter
